Migrate compile and syntax errors
